refactor: split validate_config out of config.py
config.py mixed dataclass definitions, YAML parsing, and ~540 lines of
validation. Validation (ConfigError, ConfigWarning, validate_config,
print_validation_report) now lives in src/config_validation.py; all
names are re-exported from src.config so existing imports keep working.
config.py drops from 1006 to 476 lines.
